Removed redux completely and added useQuery for logout
