+++ package.json
@@ -18,12 +18,9 @@
     "react-icons": "^4.1.0",
     "react-login-github": "^1.0.7",
     "react-query": "^3.17.0",
-    "react-redux": "^7.2.2",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
     "socket.io-client": "^3.1.1",
     "web-vitals": "^0.2.4",
     "zustand": "^3.5.2"
