--- src/actions/types.js
@@ -1,39 +1,0 @@
-// Login Actions
-export const LOGIN_LOADING = 'LOGIN_LOADING';
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
-export const LOGIN_FAIL = 'LOGIN_FAIL';
-
-// Signup Actions
-export const SIGNUP_LOADING = 'SIGNUP_LOADING';
-export const SIGNUP_SUCCESS = 'SIGNUP_SUCCESS';
-export const SIGNUP_FAIL = 'SIGNUP_FAIL';
-
-// PreCheck Actions
-export const PRECHECK_LOADING = 'PRECHECK_LOADING';
-export const PRECHECK_SUCCESS = 'PRECHECK_SUCCESS';
-export const PRECHECK_FAIL = 'PRECHECK_FAIL';
-
-// Logout Actions
-export const LOGOUT_LOADING = 'LOGOUT_LOADING';
-export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-export const LOGOUT_FAIL = 'LOGOUT_FAIL';
-
-// Socket Actions
-export const SOCKET_CONN_SUCCESS = 'SOCKET_CONN_SUCCESS';
-export const SOCKET_CONN_FAIL = 'SOCKET_CONN_FAIL';
-
-// Room Actions
-export const CREATE_ROOM_LOADING = 'CREATE_ROOM_LOADING';
-export const CREATE_ROOM_SUCCESS = 'CREATE_ROOM_SUCCESS';
-export const CREATE_ROOM_FAIL = 'CREATE_ROOM_FAIL';
-
-export const JOIN_ROOM_LOADING = 'JOIN_ROOM_LOADING';
-export const JOIN_ROOM_SUCCESS = 'JOIN_ROOM_SUCCESS';
-export const JOIN_ROOM_FAIL = 'JOIN_ROOM_FAIL';
-
-export const NEW_ROOM_MESSAGE = 'NEW_ROOM_MESSAGE';
-
-export const NEW_ROOM_MEMBER = 'NEW_ROOM_MEMBER';
-
-// Reset Action
-export const ACTION_RESET = 'ACTION_RESET';
