+++ src/components/userProfileButton/UserProfileButton.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Image } from '@chakra-ui/react';
 import getUserData from '../../utils/getUserData';
 import {
@@ -11,76 +11,11 @@ import {
   Button,
   Text,
   Flex,
-  useToast,
 } from '@chakra-ui/react';
-import { connect } from 'react-redux';
-import { logoutUser, userActionReset } from '../../actions/userActions';
-import { useHistory } from 'react-router-dom';
-import { LOGOUT, ERROR } from '../../utils/constants';
+import Logout from './Logout';
 
-const UserProfileButton = ({ logoutUser, userActionReset, userData }) => {
+const UserProfileButton = () => {
   const user = getUserData();
-  const history = useHistory();
-  const toast = useToast();
-
-  // Handle logout button click
-  const handleLogout = () => {
-    logoutUser(history);
-  };
-
-  // Logout user error handling
-  useEffect(() => {
-    if (
-      userData.logoutData.error &&
-      userData.logoutData.error.payload !== undefined
-    ) {
-      switch (userData.logoutData.error) {
-        case ERROR:
-          toast({
-            title: 'Error on Logout',
-            description: "Couldn't logout, please try again later!",
-            status: 'error',
-            position: 'top-right',
-            duration: 4000,
-            isClosable: false,
-          });
-          userActionReset();
-          break;
-        default:
-          toast({
-            title: 'Error on Logout',
-            description: "Couldn't logout, please try again later!",
-            status: 'error',
-            position: 'top-right',
-            duration: 4000,
-            isClosable: false,
-          });
-          userActionReset();
-          break;
-      }
-    } else if (userData.logoutData.error) {
-      toast({
-        title: 'Error on Logout',
-        description: userData.logoutData.error,
-        status: 'error',
-        position: 'top-right',
-        duration: 4000,
-        isClosable: false,
-      });
-      userActionReset();
-    }
-  }, [userData.logoutData.error, toast, userActionReset]);
-
-  // When logout is successfull
-  useEffect(() => {
-    if (userData.logoutData.data) {
-      if (userData.logoutData.data.payload.message === LOGOUT) {
-        userActionReset();
-        localStorage.removeItem('token');
-        history.push('/');
-      }
-    }
-  }, [userData.logoutData.data, userActionReset, history]);
 
   return (
     <Popover>
@@ -109,24 +44,11 @@ const UserProfileButton = ({ logoutUser, userActionReset, userData }) => {
           <Button colorScheme='teal' variant='ghost' w='100%'>
             Settings
           </Button>
-          <Button
-            colorScheme='red'
-            variant='ghost'
-            w='100%'
-            onClick={handleLogout}
-          >
-            Logout
-          </Button>
+          <Logout />
         </PopoverBody>
       </PopoverContent>
     </Popover>
   );
 };
 
-const mapStateToProps = (state) => ({
-  userData: state.userData,
-});
-
-export default connect(mapStateToProps, { logoutUser, userActionReset })(
-  UserProfileButton
-);
+export default UserProfileButton;
